Config file package moved to pkg folder
