+++ internal/controller/runtime/fsm/runtime_fsm_configure_oidc_test.go
@@ -2,12 +2,12 @@ package fsm
 
 import (
 	"context"
+	"github.com/kyma-project/infrastructure-manager/pkg/config"
 	"testing"
 
 	gardener "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	authenticationv1alpha1 "github.com/gardener/oidc-webhook-authenticator/apis/authentication/v1alpha1"
 	imv1 "github.com/kyma-project/infrastructure-manager/api/v1"
-	"github.com/kyma-project/infrastructure-manager/internal/config"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
