+++ internal/controller/runtime/fsm/runtime_fsm_patch_shoot.go
@@ -2,10 +2,10 @@ package fsm
 
 import (
 	"context"
+	"github.com/kyma-project/infrastructure-manager/pkg/config"
 
 	gardener "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	imv1 "github.com/kyma-project/infrastructure-manager/api/v1"
-	"github.com/kyma-project/infrastructure-manager/internal/config"
 	gardener_shoot "github.com/kyma-project/infrastructure-manager/pkg/gardener/shoot"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/utils/ptr"
