+++ internal/controller/runtime/fsm/runtime_fsm_persist_shoot_test.go
@@ -3,12 +3,12 @@ package fsm
 import (
 	"bytes"
 	"context"
+	"github.com/kyma-project/infrastructure-manager/pkg/config"
 	"github.com/stretchr/testify/mock"
 	"io"
 	"time"
 
 	gardener "github.com/gardener/gardener/pkg/apis/core/v1beta1"
-	"github.com/kyma-project/infrastructure-manager/internal/config"
 	"github.com/kyma-project/infrastructure-manager/internal/controller/metrics/mocks"
 	"github.com/kyma-project/infrastructure-manager/internal/controller/runtime/fsm/testing"
 	. "github.com/onsi/ginkgo/v2" //nolint:revive
