+++ pkg/gardener/shoot/extender/oidc.go
@@ -3,7 +3,7 @@ package extender
 import (
 	gardener "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	imv1 "github.com/kyma-project/infrastructure-manager/api/v1"
-	"github.com/kyma-project/infrastructure-manager/internal/config"
+	"github.com/kyma-project/infrastructure-manager/pkg/config"
 	"k8s.io/utils/ptr"
 )
