+++ pkg/gardener/shoot/extender/oidc_test.go
@@ -1,11 +1,11 @@
 package extender
 
 import (
+	"github.com/kyma-project/infrastructure-manager/pkg/config"
 	"testing"
 
 	gardener "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	imv1 "github.com/kyma-project/infrastructure-manager/api/v1"
-	"github.com/kyma-project/infrastructure-manager/internal/config"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
